% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_offset_variable_survival}
\alias{create_offset_variable_survival}
\title{Helper functions}
\usage{
create_offset_variable_survival(nodes, delta, time_to_event)
}
\description{
This script contains the utils functions that are used in the package.
}
