% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_response_variable_c_risks}
\alias{create_response_variable_c_risks}
\title{Helper functions}
\usage{
create_response_variable_c_risks(nodes, time_to_event, delta, event_type)
}
\description{
This script contains the utils functions that are used in the package.
}
