% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreProcessing.R
\name{data_pre_processing}
\alias{data_pre_processing}
\title{Individual Data Pre-Processing}
\usage{
data_pre_processing(data, id, status, event_time, nodes = NULL)
}
\arguments{
\item{data}{\code{data.table}, input data to be pre-processed.}

\item{id}{\code{character}, identifier column.}

\item{status}{\code{character}, status column.}

\item{nodes}{\code{numeric}, time grid to construct the piece-wise constant model.}
}
\value{
\code{data.table} containing the pre-processed data.
}
\description{
This function pre-processes the data for the application of our models.
}
