% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreProcessing.R
\name{DataPreProcessing}
\alias{DataPreProcessing}
\title{Individual Data Pre-Processing}
\usage{
DataPreProcessing(
  data,
  id,
  status,
  setting = "survival",
  covariates,
  treatment,
  event
)
}
\description{
This function pre-processes the data for the application of our models.
}
