% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishNet.R
\name{fishNet}
\alias{fishNet}
\title{fit poisson regression model}
\usage{
fishNet(
  data,
  event_time,
  status,
  covariates,
  alpha,
  lambda_grid,
  penalise_nodes = FALSE,
  number_of_nodes = 20,
  nfold = 20
)
}
\arguments{
\item{data}{data including the outcome variables and the covariates}

\item{event_time}{character. event time variable}

\item{status}{character. event type.}

\item{covariates}{character vector of covariate names}

\item{alpha}{alpha mixing parameter (1=LASSO, 0=ridge, elastic net else)}

\item{lambda_grid}{vector of values for lambda search}

\item{penalise_nodes}{if TRUE penalize also baseline hazard parameters}

\item{number_of_nodes}{number of baseline hazard parameters}

\item{nfold}{number of folds for super learner}
}
\value{
fitted object
}
\description{
Fit poisson regression model
}
\details{
via Superlearner
}
\examples{
covariates <- c("sex","age","diabetes_duration","value_SBP","value_LDL",
    "value_HBA1C","value_Smoking","value_Motion","value_Albuminuria","eGFR")
d <- simulateStenoT1(n = 300,beta_age_rate_cvd = .1,beta_age_rate_death = 0,
                     keep = c("time","event",covariates))
fit <- fishNet(data = d,event_time = "time",status = "event",
 covariates = covariates,alpha = 0,lambda = seq(.001,.9,.0002),
 penalise_nodes = FALSE,number_of_nodes=20,nfold = 20)

}
\author{
Thomas A. Gerds \href{mailto:tag@biostat.ku.dk}{tag@biostat.ku.dk}
}
