% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{learner_glmnet}
\alias{learner_glmnet}
\title{\code{glmnet} learner}
\usage{
learner_glmnet(formula, cross_validation = FALSE, learner_label = NULL, ...)
}
\arguments{
\item{cross_validation}{\code{logical}, whether the learner hyper-parameters need to be learnt for each data split.}

\item{learner_label}{\code{character} and optional, user-definer label for the learner.}

\item{...}{any additional argument to pass to \code{glmnet} or \code{cv.glmnet}.}

\item{data}{\code{data.table}, input data to be pre-processed.}

\item{id}{\code{character}, identifier column.}
}
\value{
\code{data.table} containing the pre-processed data.
}
\description{
Learner that uses the \code{glmnet} algorithm.
}
