% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\docType{class}
\name{Learner_xgboost-class}
\alias{Learner_xgboost-class}
\alias{Learner_xgboost}
\title{Piece-wise constant hazard via \code{xgboost}}
\description{
Piece-wise constant hazard via \code{xgboost}
}

