% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_deviance.R
\name{poisson_deviance}
\alias{poisson_deviance}
\alias{poisson_deviance.poisson_superlearner}
\title{Compute Poisson deviance for Poisson Super Learner objects}
\usage{
poisson_deviance(object, ...)

\method{poisson_deviance}{poisson_superlearner}(object, newdata, ...)
}
\arguments{
\item{object}{A fitted object of class \code{poisson_superlearner} returned by
\code{\link[=Superlearner]{Superlearner()}}.}

\item{...}{Additional arguments passed to methods. Currently unused.}

\item{newdata}{A \code{data.frame} containing the data that should be used to
evaluate the Poisson deviance.}
}
\value{
A list with two elements: \code{learners}, a \code{data.table} containing the
Poisson deviance for each base learner, and \code{metalearner}, a \code{data.table}
with the deviance for the meta learner. If the super learner was trained
with a single learner, the \code{metalearner} entry is \code{NULL}.
}
\description{
This method evaluates the out-of-sample Poisson deviance of the fitted
learners that form the ensemble as well as the meta learner on new data.
}
