% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fishNet.R
\name{print.fishNet}
\alias{print.fishNet}
\title{print fishNet objects}
\usage{
\method{print}{fishNet}(x, ...)
}
\arguments{
\item{x}{object obtained with \link[tmlensemble]{fishNet}.}

\item{...}{passed to print}
}
\value{
object
}
\description{
print fishNet objects
}
\details{
show beta coefficients
}
\author{
Thomas A. Gerds \href{mailto:tag@biostat.ku.dk}{tag@biostat.ku.dk}
}
