% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stratified_sampling}
\alias{stratified_sampling}
\title{Helper functions}
\usage{
stratified_sampling(dt, id, status, nfold)
}
\description{
This script contains the utils functions that are used in the package.
}
