% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearnercv.R
\name{Superlearner.cv}
\alias{Superlearner.cv}
\title{K-fold Cross-Validation Poisson Super Learner}
\usage{
Superlearner.cv(
  data,
  id = "id",
  stratified_k_fold = FALSE,
  start_time = NULL,
  end_time = NULL,
  status = "status",
  event_time = NULL,
  learners,
  n_cross_validation_folds = 5,
  nodes = NULL,
  meta_learner_algorithm = "glmnet",
  add_nodes_metalearner = TRUE,
  add_intercept_metalearner = TRUE,
  matrix_transformation = FALSE,
  penalise_nodes_metalearner = TRUE,
  variable_transformation = NULL,
  nfold = 3
)
}
\arguments{
\item{data}{\code{data.frame}, input data to be pre-processed.}

\item{id}{\code{character}, identifier column.}

\item{status}{\code{character}, status column.}

\item{nodes}{\code{numeric}, time grid to construct the piece-wise constant model.}
}
\value{
superlearner
}
\description{
This function implements the Poisson Super Learner.
}
