% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmlensemble-package.R
\docType{package}
\name{tmlensemble-package}
\alias{tmlensemble}
\alias{tmlensemble-package}
\title{tmlensemble: Poisson Super-learner}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
