% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.poisson_superlearner}
\alias{coef.poisson_superlearner}
\title{Coef method for poisson_superlearner objects}
\usage{
\method{coef}{poisson_superlearner}(object, cause = NULL, ...)
}
\description{
Coef method for poisson_superlearner objects
}
