% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\docType{class}
\name{Learner_xgboost-class}
\alias{Learner_xgboost-class}
\alias{Learner_xgboost}
\title{Script containing our learners}
\description{
This script contains the learners available in the package.
}

