% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonsuperlearner-package.R
\docType{package}
\name{poissonsuperlearner-package}
\alias{poissonsuperlearner}
\alias{poissonsuperlearner-package}
\title{poissonsuperlearner: Poisson Super Learner}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Gabriele Pittarello \email{gabriele.pittarello@sund.ku.dk}

}
\keyword{internal}
