% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonsuperlearner-package.R
\docType{package}
\name{poissonsuperlearner-package}
\alias{poissonsuperlearner}
\alias{poissonsuperlearner-package}
\title{poissonsuperlearner: Poisson Super Learner}
\description{
Provides an implementation of piece-wise constant hazard models for time-to-event analysis of survival and competing risks data. The package includes an ensemble approach, the Poisson Superlearner, for flexible hazard estimation. It enables estimation of survival functions and risk predictions.
}
\author{
\strong{Maintainer}: Gabriele Pittarello \email{gabriele.pittarello@sund.ku.dk}

Authors:
\itemize{
  \item Helene Rytgaard \email{hely@sund.ku.dk}
  \item Thomas Gerds \email{tag@biostat.ku.dk}
}

}
\keyword{internal}
