% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonsuperlearner-package.R
\docType{package}
\name{poissonsuperlearner-package}
\alias{poissonsuperlearner}
\alias{poissonsuperlearner-package}
\title{poissonsuperlearner: Poisson Super Learner}
\description{
Provides tools for fitting piece-wise constant hazard models for survival and competing risks data, including ensemble hazard estimation via the Super Learner framework. The package supports estimation of survival functions and absolute risk predictions from fitted cause-specific hazard models. For the Super Learner framework see van der Laan, Polley and Hubbard (2007) \doi{10.2202/1544-6115.1309}.
}
\author{
\strong{Maintainer}: Gabriele Pittarello \email{gabriele.pittarello@sund.ku.dk}

Authors:
\itemize{
  \item Helene Rytgaard \email{hely@sund.ku.dk}
  \item Thomas Gerds \email{tag@biostat.ku.dk}
}

}
\keyword{internal}
