% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.poisson_superlearner}
\alias{predict.poisson_superlearner}
\title{Individual Data Pre-Processing}
\usage{
\method{predict}{poisson_superlearner}(object, newdata, type = "survival", k = 1, step = 0.5, ...)
}
\value{
predictions
}
\description{
This function predicts the survivor function or the hazard using the superlearner.
}
