% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.poisson_superlearner}
\alias{predict.poisson_superlearner}
\title{Poisson Super-Learner Predictions}
\usage{
\method{predict}{poisson_superlearner}(object, newdata, times, cause = 1, ...)
}
\arguments{
\item{object}{\code{poisson_superlearner} for absolute risk and survival function predictions.}

\item{newdata}{\code{data.frame}, new data to predict the absolute risk and the survival function for.}

\item{times}{\code{numeric}, time(s) at which to predict the absolute risk and survival function.}

\item{cause}{\code{numeric}, competing risk to predict the absolute risk and the survival function for.}
}
\value{
\code{data.table} containing for each row of \code{newdata} the \code{poisson_superlearner} predictions for the survival function and the absolute risk predictions for some \code{cause} at the given \code{times}.
}
\description{
This method computes the survival function and the absolute risk prediction of a \code{poisson_superlearner} object, based on a given data set, at given \code{times} for a given \code{cause}.
}
