% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.poisson_superlearner}
\alias{predict.poisson_superlearner}
\title{Predictions of the Poisson Super-Learner}
\usage{
\method{predict}{poisson_superlearner}(object, newdata, times, type = "survival", cause = 1, ...)
}
\value{
predictions
}
\description{
This function predicts the survivor function, the piece-wise constant hazard and the absolute risk using the superlearner.
}
