% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRisk.R
\name{predictRisk.poisson_superlearner}
\alias{predictRisk.poisson_superlearner}
\title{Absolute-risk matrix predictions for a fitted Poisson Super Learner}
\usage{
\method{predictRisk}{poisson_superlearner}(object, newdata, times, cause = 1, model = "sl", ...)
}
\arguments{
\item{object}{\code{poisson_superlearner}. Fitted object.}

\item{newdata}{\code{data.frame}. New covariate data.}

\item{times}{\code{numeric}. Prediction times.}

\item{cause}{\code{numeric(1)}. Cause index.}

\item{model}{Scalar model selector. Default is \code{"sl"}. Allowed values are the
same as in \code{\link[=predict.poisson_superlearner]{predict.poisson_superlearner()}}.}

\item{...}{Unused.}
}
\value{
\code{numeric} matrix with \code{nrow(newdata)} rows and \code{length(times)} columns.
}
\description{
S3 method compatible with \code{riskRegression::predictRisk} returning one column
per requested time.
}
