% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRisk.R
\name{predictRisk.poisson_superlearner}
\alias{predictRisk.poisson_superlearner}
\title{Individual Data Pre-Processing}
\usage{
\method{predictRisk}{poisson_superlearner}(
  object,
  newdata,
  times,
  cause = 1,
  absolute_risk_integration = "exact",
  ...
)
}
\value{
predictions
}
\description{
Extract event probabilities from fitted poisson_superlearner.
}
