% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.poisson_superlearner}
\alias{print.poisson_superlearner}
\title{Print method for poisson_superlearner objects}
\usage{
\method{print}{poisson_superlearner}(object, cause = 1, ...)
}
\description{
Print method for poisson_superlearner objects
}
