% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateStenoT1.R
\name{simulateStenoT1}
\alias{simulateStenoT1}
\title{Simulate time to CVD of hypothetical type-1 diabetes patients}
\usage{
simulateStenoT1(
  n,
  coefficient_age = 0.05,
  coefficient_LDL = 0.1,
  value_diabetis = 0.02,
  keep = NULL,
  scenario = c("alpha", "beta"),
  competing_risks = FALSE
)
}
\arguments{
\item{n}{Sample size}

\item{keep}{Optional vector of variables to keep in the data set}

\item{beta_age_rate_cvd}{Log-hazard ratio for age on the rate of CVD}

\item{beta_LDL_rate_cvd}{Log-hazard ratio for LDL on the rate of CVD}

\item{beta_diabetes_duration_rate_cvd}{Log-hazard ratio for diabetes duration on the rate of CVD}
}
\value{
data.table with baseline covariates and
time to cvd (time,event) where event has values 0 for right censored
1 for cvd and 2 for death without cvd.
}
\description{
Simulate data that are alike those behind the Steno type-1 risk engine
}
\details{
The data are computer generated and do not belong to real people.
}
\examples{
simulateStenoT1(n=3)
}
\author{
Thomas A. Gerds \href{mailto:tag@biostat.ku.dk}{tag@biostat.ku.dk}
}
