% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateStenoT1.R
\name{simulateStenoT1}
\alias{simulateStenoT1}
\title{Simulate time to CVD of hypothetical type-1 diabetes patients}
\usage{
simulateStenoT1(
  n,
  beta_sex_age = 0.105,
  beta_age_diabetes_duration = 0.461,
  beta_sex_diabetes_duration = 1.883,
  beta_diabetes_duration_value_SBP = 0.174,
  beta_age_value_SBP = 0.408,
  beta_sex_value_SBP = -5.264,
  beta_value_SBP_value_LDL = 0.0018,
  beta_diabetes_duration_value_LDL = -0.00215,
  beta_age_value_LDL = -0.00194,
  beta_sex_value_LDL = -0.02479,
  beta_value_LDL_value_HBA1C = 2.414,
  beta_value_SBP_value_HBA1C = -0.03665,
  beta_diabetes_duration_value_HBA1C = 0.1058,
  beta_age_value_HBA1C = -0.11352,
  beta_sex_value_HBA1C = 1.144,
  beta_value_AlbuminuriaMicro_log2_eGFR_young = -0.038,
  beta_value_AlbuminuriaMacro_log2_eGFR_young = 0.37,
  beta_value_HBA1C_log2_eGFR_young = -0.00038,
  beta_value_LDL_log2_eGFR_young = 0.076,
  beta_value_SBP_log2_eGFR_young = -0.00092,
  beta_diabetes_duration_log2_eGFR_young = 0.011,
  beta_age_log2_eGFR_young = 0.174,
  beta_sex_log2_eGFR_young = -0.051,
  beta_log2_eGFR_young_log2_eGFR_old = -0.998,
  beta_value_AlbuminuriaMicro_log2_eGFR_old = 0.132,
  beta_value_AlbuminuriaMacro_log2_eGFR_old = 0.627,
  beta_value_HBA1C_log2_eGFR_old = -0.00173,
  beta_value_LDL_log2_eGFR_old = -0.01687,
  beta_value_SBP_log2_eGFR_old = 0.00066,
  beta_diabetes_duration_log2_eGFR_old = 0.00232,
  beta_age_log2_eGFR_old = 0.01052,
  beta_sex_log2_eGFR_old = 0.06662,
  beta_log2_eGFR_old_value_Smoking = -0.573,
  beta_log2_eGFR_young_value_Smoking = -0.533,
  beta_value_AlbuminuriaMicro_value_Smoking = 0.418,
  beta_value_AlbuminuriaMacro_value_Smoking = 0.812,
  beta_value_HBA1C_value_Smoking = 0.026,
  beta_value_LDL_value_Smoking = 0.128,
  beta_value_SBP_value_Smoking = -0.0102,
  beta_diabetes_duration_value_Smoking = -0.00575,
  beta_age_value_Smoking = 0.0053,
  beta_sex_value_Smoking = -0.131,
  beta_value_Smoking_value_Motion = -0.311,
  beta_log2_eGFR_old_value_Motion = -0.134,
  beta_log2_eGFR_young_value_Motion = -0.086,
  beta_value_AlbuminuriaMicro_value_Motion = -0.121,
  beta_value_AlbuminuriaMacro_value_Motion = -0.335,
  beta_value_HBA1C_value_Motion = -0.01001,
  beta_value_LDL_value_Motion = -0.138,
  beta_value_SBP_value_Motion = 0.0069,
  beta_diabetes_duration_value_Motion = -0.00803,
  beta_age_value_Motion = -0.00518,
  beta_sex_value_Motion = 0.071,
  beta_value_Motion_rate_censored = -0.039,
  beta_value_Smoking_rate_censored = -0.106,
  beta_log2_eGFR_old_rate_censored = -0.137,
  beta_log2_eGFR_young_rate_censored = -0.119,
  beta_value_AlbuminuriaMicro_rate_censored = -0.033,
  beta_value_AlbuminuriaMacro_rate_censored = 0.076,
  beta_value_HBA1C_rate_censored = -0.00618,
  beta_value_LDL_rate_censored = 0.051,
  beta_value_SBP_rate_censored = -0.00757,
  beta_diabetes_duration_rate_censored = -0.01259,
  beta_age_rate_censored = -0.03,
  beta_sex_rate_censored = -0.02963,
  beta_value_Motion_rate_cvd = -0.176,
  beta_value_Smoking_rate_cvd = 0.311,
  beta_log2_eGFR_old_rate_cvd = 0.044,
  beta_log2_eGFR_young_rate_cvd = 0.48,
  beta_value_AlbuminuriaMicro_rate_cvd = 0.046,
  beta_value_AlbuminuriaMacro_rate_cvd = 0.076,
  beta_value_HBA1C_rate_cvd = 0.01437,
  beta_value_LDL_rate_cvd = 0.1,
  beta_value_SBP_rate_cvd = 0.00608,
  beta_diabetes_duration_rate_cvd = 0.02,
  beta_age_rate_cvd = 0.05,
  beta_sex_rate_cvd = -0.224,
  beta_value_Motion_rate_death = -0.411,
  beta_value_Smoking_rate_death = 0.834,
  beta_log2_eGFR_old_rate_death = 0.329,
  beta_log2_eGFR_young_rate_death = 0.313,
  beta_value_AlbuminuriaMicro_rate_death = 0.226,
  beta_value_AlbuminuriaMacro_rate_death = 0.604,
  beta_value_HBA1C_rate_death = 0.01025,
  beta_value_LDL_rate_death = 0.09,
  beta_value_SBP_rate_death = 0.00156,
  beta_diabetes_duration_rate_death = 0.06,
  beta_age_rate_death = 0.09,
  beta_sex_rate_death = -0.276,
  keep = NULL
)
}
\arguments{
\item{n}{Sample size}

\item{beta_diabetes_duration_rate_cvd}{Log-hazard ratio for diabetes duration on the rate of CVD}

\item{beta_age_rate_cvd}{Log-hazard ratio for age on the rate of CVD}

\item{keep}{Optional vector of variables to keep in the data set}

\item{beta_LDL_rate_cvd}{Log-hazard ratio for LDL on the rate of CVD}
}
\value{
data.table with baseline covariates and
time to cvd (time,event) where event has values 0 for right censored
1 for cvd and 2 for death without cvd.
}
\description{
Simulate data that are alike those behind the Steno type-1 risk engine
}
\details{
The data are computer generated and do not belong to real people.
}
\examples{
simulateStenoT1(n=3)
}
\author{
Thomas A. Gerds \href{mailto:tag@biostat.ku.dk}{tag@biostat.ku.dk}
}
