% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.base_learner}
\alias{summary.base_learner}
\title{Summary method for base_learner objects}
\usage{
\method{summary}{base_learner}(object, cause = 1, ...)
}
\description{
Summary method for base_learner objects
}
