% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.poisson_superlearner}
\alias{summary.poisson_superlearner}
\title{Summarize a fitted Poisson Super Learner object}
\usage{
\method{summary}{poisson_superlearner}(object, ...)
}
\arguments{
\item{object}{\code{poisson_superlearner}.}

\item{...}{Unused.}
}
\value{
\code{data.table} with cross-validated Poisson deviance summaries for
learners and meta-learner.
}
\description{
Summarize a fitted Poisson Super Learner object
}
