% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.poisson_superlearner}
\alias{summary.poisson_superlearner}
\title{Summary method for poisson_superlearner objects}
\usage{
\method{summary}{poisson_superlearner}(object, ...)
}
\arguments{
\item{object}{A \code{poisson_superlearner} object.}

\item{...}{Unused.}
}
\description{
Prints:
(a) the meta-learner call (with Z* renamed using data_info$learners_label),
(b) the cross-validation deviance data.table,
(c) the meta-learner coefficients (with Z* renamed using data_info$learners_label).
}
