% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner.R
\name{Superlearner}
\alias{Superlearner}
\title{Individual Data Pre-Processing}
\usage{
Superlearner(
  data,
  id = "id",
  status = "status",
  event_time = "time",
  learners,
  nodes = NULL,
  nfold = 3
)
}
\arguments{
\item{data}{\code{data.frame}, input data to be pre-processed.}

\item{id}{\code{character}, identifier column.}

\item{status}{\code{character}, status column.}

\item{nodes}{\code{numeric}, time grid to construct the piece-wise constant model.}
}
\value{
superlearner
}
\description{
This function pre-processes the data for the application of our models.
}
