% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner.R
\name{Superlearner}
\alias{Superlearner}
\title{Poisson Super Learner implementation}
\usage{
Superlearner(
  data,
  id = "id",
  stratified_k_fold = FALSE,
  start_time = NULL,
  end_time = NULL,
  status = "status",
  event_time = NULL,
  learners,
  number_of_nodes = NULL,
  nodes = NULL,
  meta_learner_algorithm = "glmnet",
  add_nodes_metalearner = TRUE,
  add_intercept_metalearner = FALSE,
  matrix_transformation = FALSE,
  penalise_nodes_metalearner = FALSE,
  variable_transformation = NULL,
  nested_cross_validation_meta_learner = TRUE,
  nfold = 3,
  ...
)
}
\arguments{
\item{data}{\code{data.frame}, input data to be pre-processed.}

\item{id}{\code{character}, identifier column.}

\item{status}{\code{character}, status column.}

\item{nodes}{\code{numeric}, time grid to construct the piece-wise constant model.}
}
\value{
superlearner
}
\description{
This function implements the Poisson SuperLearner.
}
