% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner.R
\name{Superlearner}
\alias{Superlearner}
\title{Poisson Super Learner}
\usage{
Superlearner(
  data,
  id = "id",
  status = "status",
  event_time = NULL,
  learners,
  number_of_nodes = NULL,
  nodes = NULL,
  min_depth = NULL,
  meta_learner_algorithms = c("glm", "glmnet"),
  variable_transformation = NULL,
  nfold = 3,
  ...
)
}
\arguments{
\item{data}{\code{data.frame}, input data to be pre-processed.}

\item{id}{\code{character}, identifier column.}

\item{status}{\code{character}, status column.}

\item{event_time}{\code{character}, time-to-event column in competing risks and survival applications.}

\item{learners}{\code{list}, list of learners to include in the ensemble. If only one learner is provided, the learner is applied to the full data.}

\item{number_of_nodes}{\code{numeric}, number of time points sampled from the observed time to construct the nodes. Alternative to \code{nodes}, if both NULL we take all the observed time points as nodes.}

\item{nodes}{\code{numeric}, time grid to construct the piece-wise constant model. Alternative to \code{number_of_nodes}, if both NULL we take all the observed time points as nodes.}

\item{variable_transformation}{\code{list}, variable transformation(s) to apply to the data. Each element of the list is a character \code{"new_variable ~ some_function(variable_in_the_data)"}}

\item{nfold}{\code{numeric}, number of V-folds to construct the ensemble.}
}
\value{
A \code{poisson_superlearner} object contains the following output
\itemize{
\item{\code{learners}: \code{list} containing the learners.}
\item{\code{metalearner}: \code{numeric} Training negative log likelihood.}
\item{\code{superlearner}:  \code{numeric} Validation  negative log likelihood. Not available for COX.}
\item{\code{meta_learner_cross_validation}:  \code{data.table} containing the average Poisson Deviance evaluated on the V-Folds for the learners and the meta-learner.}
\item{\code{data_info}: \code{list} containing the following meta-data:}
\itemize{
\item{\code{id}: \code{character}, name of the covariate in the data that contains the id variable.}
\item{\code{status}: \code{character}, name of the covariate in the data that contains the status variable.}
\item{\code{event_time}: \code{character}, for competing risks or survival data it contains the name of the covariate containing the event time.}
\item{\code{nfold}: \code{integer} denoting the number of V-Folds.}
\item{\code{maximum_followup}: \code{numeric} denoting the maximum follow-up time observed.}
\item{\code{n_crisks}, \code{integer} denoting the number of competing risks.}
\item{\code{variable_transformation}: \code{list} or \code{array}, containing the data variable transformations implemented. }
\item{\code{interval_data_type}: \code{logical}, \code{TRUE} for interval data. }
}
\item{\code{hazard_model}: \code{string} chosen hazard model (COX, NN or XGB)}
\item{\code{IndividualDataPP}: starting \code{IndividualDataPP} object.}
}
}
\description{
This class is used for estimating the Poisson SuperLearner.
}
