% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner.R
\name{Superlearner}
\alias{Superlearner}
\title{Poisson Super Learner implementation}
\usage{
Superlearner(
  data,
  id = "id",
  stratified_k_fold = FALSE,
  start_time = NULL,
  end_time = NULL,
  status = "status",
  event_time = NULL,
  learners,
  number_of_nodes = NULL,
  nodes = NULL,
  min_depth = 2,
  meta_learner_algorithms = c("glm", "glmnet"),
  variable_transformation = NULL,
  nfold = 3,
  ...
)
}
\arguments{
\item{data}{\code{data.frame}, input data to be pre-processed.}

\item{id}{\code{character}, identifier column.}

\item{status}{\code{character}, status column.}

\item{nodes}{\code{numeric}, time grid to construct the piece-wise constant model.}
}
\value{
superlearner
}
\description{
This function implements the Poisson SuperLearner.
}
